Add server leaderboard at the end of the game

Closes #54

See merge request cse1105/2021-2022/team-repositories/oopp-group-53/repository-template!75

+++ commons/src/main/java/commons/GameState.java
@@ -52,7 +52,7 @@ public class GameState {
         this.emotes = game.emotes;
 
         for (Player lobbyPlayer : game.players) {
-            this.leaderboard.add(new LeaderboardEntry(lobbyPlayer.username, (int) lobbyPlayer.score));
+            this.leaderboard.add(new LeaderboardEntry(lobbyPlayer.username, lobbyPlayer.score));
         }
         Collections.sort(this.leaderboard);
     }
